名前空間の明示と、バージョン確認の無効化(UniVRM0.120対応)

+++ EVMC4U/class/EVMC4Ustructs.cs
@@ -74,7 +74,7 @@ namespace EVMC4U {
     public struct MidiCCButton
     {
         public int knob;
-        public int active;
+        public float active;
     }
 
     public enum CalibrationState
